Moved PythonFunction#D classes into .autowrap, they are a utility class associated with autowrap. Made cimports absolute as relative cimports have been flaky in the past with older cython versions.
