+++ raysect/core/math/function/function1d/autowrap.pxd
@@ -29,6 +29,9 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from .base cimport Function1D
+from raysect.core.math.function.function1d.base cimport Function1D
+
+cdef class PythonFunction1D(Function1D):
+    cdef public object function
 
 cdef Function1D autowrap_function1d(object function)
