+++ raysect/core/math/function/function1d/base.pxd
@@ -33,10 +33,6 @@ cdef class Function1D:
     cdef double evaluate(self, double x) except? -1e999
 
 
-cdef class PythonFunction1D(Function1D):
-    cdef public object function
-
-
 cdef class AddFunction1D(Function1D):
     cdef Function1D _function1, _function2
