+++ raysect/core/math/function/function1d/constant.pxd
@@ -29,7 +29,7 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from .base cimport Function1D
+from raysect.core.math.function.function1d.base cimport Function1D
 
 
 cdef class Constant1D(Function1D):
