+++ raysect/core/math/function/function3d/autowrap.pxd
@@ -29,6 +29,9 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from .base cimport Function3D
+from raysect.core.math.function.function3d.base cimport Function3D
+
+cdef class PythonFunction3D(Function3D):
+    cdef public object function
 
 cdef Function3D autowrap_function3d(object function)
